Hirocoin ReVIVe v0.8.8.0
Patched up with a halving bypass for June of next year (2017) then a
hardcap was implemented to take place on December 18th 2019 at which
point the coin will reach very close approximate cap (due to volatile
payout based upon difficulty fluctuation exact 672M cannot be garaunteed
however the cap will be within +/- 10% of the original cap which is
acceptable when considering the other dev hardforked the chain into this
payout structure, albeit not a bad concept this makes it exceedingly
difficult the correctly configure the chain to mineout and run on
minimal payout and hash afterwords as a means of longevity and no longer
funding.

Thank you Hirocoin for allowing us the honor and opportunity to work
with you and creating a better Hirocoin. We hope that this patch brings
you ease of use and joy as it will help gauranty the chain's stability
and health.

Hirocoin Stats and Hardcarp Calc

116M mined by block 290-K
154M mined by block 840-K (Halving Occured) ~270M
----------------------106,644 Blocks moved since Halving Occured
----------------------733,356 Blocks til next Halving
14,930,160 mined by current block (BLOCK 946,644)
284,930,160 Currently Mined total: 946644-BLOCKS

Approx current total ~284930160 /~292M (out of) 672M (due to payout
structure)
300M used for calculation simplicity

102,669,840 more coins will be mined by next Halving
total supply will be at that time:  ~387,600,000-HIRO
This will be on BLOCK:1,680,000 or June 30, 2017
Today is February 7, 2016
so that means that 509 days from today would be June 30, 2017

MINEOUT WILL OCCUR AFTER 2,031,428 MORE BLOCKS
TOTALLING 2,978,072 blocks at which point the hardcap will occur
(3.86 Years from now)
DATE THIS WILL OCCUR: December 18, 2019
* Rounded hardcap to block 2,978,250 to allow for the volatile payout.
[Today is February 7, 2016
so that means that 1410 days from today
would be December 18, 2019.]